+++ routes/api/comment-votes.js
@@ -27,17 +27,29 @@ router.post(
 	Auth.userCheck(),
 	CommentVoteAuth.verifyNonExistance(),
 	async (req, res) => {
-		const commentVotes = await Collections.loadCommentVotesCollection()
-		await commentVotes.insertOne({
-			createdAt: new Date(),
-			comment_id: req.body.comment_id,
-			block_id: req.body.block_id,
-			user_id: req.body.user_id,
-			email: req.body.email,
-			username: req.body.username,
-		})
+		if (req.body.user_id == req.decoded._id) {
+			const commentVotes = await Collections.loadCommentVotesCollection()
+			await commentVotes.insertOne({
+				createdAt: new Date(),
+				comment_id: req.body.comment_id,
+				block_id: req.body.block_id,
+				user_id: req.body.user_id,
+				email: req.body.email,
+				username: req.body.username,
+			})
 
-		res.status(201).send()
+			res.status(201).send({
+				auth: true,
+				error: 'Created commentVote'
+			})
+
+		}
+		else {
+			return res.status(401).send({
+				auth: false,
+				error: 'Bro you cant create commentVote for someone else!'
+			})
+		}
 	}
 )
 
@@ -47,13 +59,24 @@ router.delete(
 	'/delete/:user_id/:comment_id',
 	Auth.userCheck(),
 	async (req, res) => {
-		const commentVotes = await Collections.loadCommentVotesCollection()
-		await commentVotes.deleteMany({
-			comment_id: req.params.comment_id,
-			user_id: req.params.user_id,
-		})
+		if (req.params.user_id == req.decoded._id) {
+			const commentVotes = await Collections.loadCommentVotesCollection()
+			await commentVotes.deleteMany({
+				comment_id: req.params.comment_id,
+				user_id: req.params.user_id,
+			})
 
-		res.status(200).send()
+			res.status(200).send({
+				auth: true,
+				message: 'Deleted commentVote'
+			})
+		}
+		else {
+			return res.status(401).send({
+				auth: false,
+				error: 'Bro you cant delete commentVote for someone else!'
+			})
+		}
 	}
 )
 
@@ -68,7 +91,10 @@ router.delete(
 		comment_id: req.params.comment_id,
 	})
 
-	res.status(200).send()
+	res.status(200).send({
+		auth: true,
+		message: 'Deleted ALL commentVotes for a specific comment'
+	})
 })
